+++ components/home/schedule-list-view.tsx
@@ -13,8 +13,7 @@ import type { SchPostType } from "@/lib/validations/schedule-types";
 import { Caption, Micro, SectionLabel } from "@/components/common/typography";
 
 import type { CalendarRace } from "./mini-calendar";
-import type { FilterType } from "./mini-calendar";
-import { matchesFilter } from "./mini-calendar";
+import { type FilterType, matchesFilter } from "./schedule-filter";
 
 type MonthData = {
   monthKey: string; // "YYYY-MM"
